src.ts/_tests/convertToSpasm.test.ts
import {
  validDmpEventSignedClosedWithInvalidSignature,
  validDmpEventSignedClosedWithInvalidSignedString,
  validDmpEventSignedClosedWithInvalidSigner,
  validNostrSpasmEventSignedOpenedWithInvalidContent,
  validNostrSpasmEventSignedOpenedWithInvalidSignature,
  validNostrSpasmEventSignedOpenedWithInvalidSigner,
  validPostWithDmpEventSignedClosed,
  validPostWithDmpEventSignedClosedConvertedToSpasmEventEnvelopeV2,
  validPostWithDmpEventSignedClosedConvertedToSpasmV2,

  // DMP
  validDmpEvent,
  validDmpEventSignedClosed,
  validDmpEventSignedOpened,
  validSpasmDmpEventSignedClosedV0,

  // DMP converted to SpasmEventV2
  validDmpEventConvertedToSpasmEventV2,
  validDmpEventSignedClosedConvertedToSpasmV2,
  validDmpEventSignedOpenedConvertedToSpasmV2,
  validSpasmDmpEventSignedClosedV0ConvertedToSpasmV2,

  // Nostr
  validNostrEvent,
  validNostrSpasmEvent,
  validNostrEventSignedOpened,
  validNostrSpasmEventSignedOpened,
  validSpasmNostrEventSignedOpenedV0,
  validSpasmNostrSpasmEventSignedOpenedV0,

  // Nostr converted to Spasm V2
  validNostrEventConvertedToSpasmV2,
  validNostrSpasmEventConvertedToSpasmV2,
  validNostrEventSignedOpenedConvertedToSpasmV2,
  validNostrSpasmEventSignedOpenedConvertedToSpasmV2,
  validSpasmNostrEventSignedOpenedV0ConvertedToSpasmV2,
  validSpasmNostrSpasmEventSignedOpenedV0ConvertedToSpasmV2,

  // RSS items V2
  validSpasmEventRssItemV0,
  validSpasmEventRssItemV0ConvertedToSpasmV2,
  validPostWithRssItemSpecialChars,
  validPostWithRssItemSpecialCharsConvertedToSpasmEventV2,
  validPostWithRssItemTitleHasSpecialChars,
  validPostWithRssItemTitleHasSpecialCharsConvertedToSpasmEventV2,
  validNostrSpasmEventSpasmV0WithInvalidHtmlTags,
  validRssItemWithEmoji,
  validRssItemWithEmojiConvertedToSpasmEvent2,
  // validPostWithNostrReplyToDmpEvent,
  validPostWithNostrReplyToDmpEventConvertedToSpasmV2WithSpasmParentEvent,
  validDmpEventSignedClosedConvertedToSpasmV2WithTwoChildren,
  validPostWithNostrReplyToDmpEventConvertedToSpasmV2,
  validDmpEventSignedClosedConvertedToSpasmEventEnvelopeWithTreeV2WithTwoChildren
} from "./_events-data.js"
import { convertManyToSpasm, convertToSpasm } from "./../convert/convertToSpasm.js"
import {
  convertManyToSpasmEventEnvelope,
  convertToSpasmEventEnvelope
} from "./../convert/convertToSpasmEventEnvelope.js"
import {
  convertManyToSpasmEventEnvelopeWithTree,
  convertToSpasmEventEnvelopeWithTree
} from "../convert/convertToSpasmEventEnvelopeWithTree.js";
import {
  convertManyToSpasmEventDatabase,
  convertToSpasmEventDatabase
} from "../convert/convertToSpasmEventDatabase.js";
import {
  CustomConvertToSpasmConfig,
  SpasmEventV2,
  UnknownEventV2
} from "../types/interfaces.js";
import {copyOf} from "../utils/utils.js";

describe("convertToSpasm tests", () => {
  test("should return true if true", () => {
    expect(true).toBe(true);
  });
});

// convertToSpasm() for DMP events
describe("convertToSpasm() tests for DMP events", () => {
  // DmpEvent to V2
  test("should convert validDmpEvent to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validDmpEvent));
    const output = JSON.parse(JSON.stringify(validDmpEventConvertedToSpasmEventV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // DmpEventSignedClosed to V2
  test("should convert validDmpEventSignedClosed to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validDmpEventSignedClosed));
    const output = JSON.parse(JSON.stringify(validDmpEventSignedClosedConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // DmpEventSignedOpened to V2
  test("should convert validDmpEventSignedOpened to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validDmpEventSignedOpened));
    const output = JSON.parse(JSON.stringify(validDmpEventSignedOpenedConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // SpasmDmpEventSignedClosedV0 to V2
  // validPostWithDmpEventSignedClosed - old name
  // validSpasmDmpEventSignedClosedV0 - new name
  test("should convert validSpasmDmpEventSignedClosedV0 to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validSpasmDmpEventSignedClosedV0));
    const output = JSON.parse(JSON.stringify(validSpasmDmpEventSignedClosedV0ConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  test("should return null if signed Dmp events have invalid signatures", () => {
    const inputValid = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signature: "0xbd934a01dc3bd9bb183bda807d35e61accf7396c527b8a3d029c20c00b294cf029997be953772da32483b077eea856e6bafcae7a2aff95ae572af25dd3e204a71b",
      signer: "0xf8553015220a857eda377a1e903c9e5afb3ac2fa"
    };
    const inputValid1 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedOpened)),
      signature: "0xbd934a01dc3bd9bb183bda807d35e61accf7396c527b8a3d029c20c00b294cf029997be953772da32483b077eea856e6bafcae7a2aff95ae572af25dd3e204a71b",
      signer: "0xf8553015220a857eda377a1e903c9e5afb3ac2fa"
    };
    const inputInvalid1 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signature: "0xbd934a01dc3bd9bb183bda807d35e61accf7396c527b8a3d029c20c00b294cf029997be953772da32483b077eea856e6bafcae7a2aff95ae572af25dd3e204a71c"
    };
    const inputInvalid2 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signature: "dmp-invalid-signature"
    };
    const inputInvalid3 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signer: "0xf8553015220a857eda377a1e903c9e5afb3ac2fb"
    };
    const inputInvalid4 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signer: "dmp-invalid-signer"
    };
    // const inputInvalid5 = {
    //   ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
    //   signature: ""
    // };
    const inputInvalid6 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signer: ""
    };
    const inputInvalid7 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
      signature: "dmp-invalid-signature",
      signer: "dmp-invalid-signer"
    };
    // const inputInvalid8 = {
    //   ...JSON.parse(JSON.stringify(validDmpEventSignedClosed)),
    //   signature: "",
    //   signer: ""
    // };
    const inputInvalid9 = {
      ...JSON.parse(JSON.stringify(validDmpEventSignedOpened)),
      signature: "dmp-invalid-signature",
      signer: "dmp-invalid-signer"
    };
    const outputValid = JSON.parse(JSON.stringify(validDmpEventSignedClosedConvertedToSpasmV2));
    const outputValid1 = JSON.parse(JSON.stringify(validDmpEventSignedOpenedConvertedToSpasmV2));

    expect(convertToSpasm(inputValid)).toEqual(outputValid);
    expect(convertToSpasm(inputValid1)).toEqual(outputValid1);
    expect(convertToSpasm(inputInvalid1)).toEqual(null);
    expect(convertToSpasm(inputInvalid2)).toEqual(null);
    expect(convertToSpasm(inputInvalid3)).toEqual(null);
    expect(convertToSpasm(inputInvalid4)).toEqual(null);
    // expect(convertToSpasm(inputInvalid5)).toEqual(null);
    expect(convertToSpasm(inputInvalid6)).toEqual(null);
    expect(convertToSpasm(inputInvalid7)).toEqual(null);
    // expect(convertToSpasm(inputInvalid8)).toEqual(null);
    expect(convertToSpasm(inputInvalid9)).toEqual(null);
  });
});

describe("convertMany... tests for different events", () => {
  test("should convert DMP events to SpasmEventV2", () => {
    const dmpInput = JSON.parse(JSON.stringify(validDmpEvent));
    const dmpOutput = JSON.parse(JSON.stringify(validDmpEventConvertedToSpasmEventV2));
    expect(convertManyToSpasm([
      copyOf(dmpInput), copyOf(dmpInput), copyOf(dmpInput)
    ])).toEqual([
      copyOf(dmpOutput), copyOf(dmpOutput), copyOf(dmpOutput)
    ]);
  });
  test("should convert different DMP and Nostr events to SpasmEventV2", () => {
    const dmpInput = JSON.parse(JSON.stringify(validDmpEvent));
    const nostrInput = JSON.parse(JSON.stringify(validNostrEventSignedOpened));
    const dmpOutput = JSON.parse(JSON.stringify(validDmpEventConvertedToSpasmEventV2));
    const nostrOutput = JSON.parse(JSON.stringify(validNostrEventSignedOpenedConvertedToSpasmV2));
    expect(convertManyToSpasm([
      copyOf(dmpInput), copyOf(nostrInput), copyOf(dmpInput)
    ])).toEqual([
      copyOf(dmpOutput), copyOf(nostrOutput), copyOf(dmpOutput)
    ]);
  });
  test("should convert DMP events to SpasmEventEnvelopeV2", () => {
    const dmpInput = JSON.parse(JSON.stringify(validPostWithDmpEventSignedClosed));
    const dmpOutput = JSON.parse(JSON.stringify(validPostWithDmpEventSignedClosedConvertedToSpasmEventEnvelopeV2));
    expect(convertManyToSpasmEventEnvelope([
      copyOf(dmpInput), copyOf(dmpInput), copyOf(dmpInput)
    ])).toEqual([
      copyOf(dmpOutput), copyOf(dmpOutput), copyOf(dmpOutput)
    ]);
  });
  test("should convert different DMP and Nostr events to EnvelopeWithTree and then to SpasmEventV2", () => {
    const dmpInput = JSON.parse(JSON.stringify(validDmpEvent));
    const nostrInput = JSON.parse(JSON.stringify(validNostrEventSignedOpened));
    const dmpOutput = JSON.parse(JSON.stringify(validDmpEventConvertedToSpasmEventV2));
    const nostrOutput = JSON.parse(JSON.stringify(validNostrEventSignedOpenedConvertedToSpasmV2));
    const envelopesWithTree = convertManyToSpasmEventEnvelopeWithTree([
      dmpInput, nostrInput
    ])
    expect(convertManyToSpasm(envelopesWithTree!))
      .toEqual([copyOf(dmpOutput), copyOf(nostrOutput)]);
  });
  test("should convert different DMP and Nostr events to SpasmEventDatabaseV2 and then to SpasmEventV2", () => {
    const dmpInput = JSON.parse(JSON.stringify(validDmpEvent));
    const nostrInput = JSON.parse(JSON.stringify(validNostrEventSignedOpened));
    const dmpOutput = JSON.parse(JSON.stringify(validDmpEventConvertedToSpasmEventV2));
    const nostrOutput = JSON.parse(JSON.stringify(validNostrEventSignedOpenedConvertedToSpasmV2));
    const databaseEvents = convertManyToSpasmEventDatabase([
      dmpInput, nostrInput
    ])
    expect(convertManyToSpasm(databaseEvents!))
      .toEqual([copyOf(dmpOutput), copyOf(nostrOutput)]);
  });
});

// convertToSpasm() for Nostr events
describe("convertToSpasm() tests for Nostr events", () => {
  // NostrEvent to V2
  test("should convert validNostrEvent to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validNostrEvent));
    const output = JSON.parse(JSON.stringify(validNostrEventConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // NostrSpasmEvent to V2
  test("should convert validNostrSpasmEvent to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validNostrSpasmEvent));
    const output = JSON.parse(JSON.stringify(validNostrSpasmEventConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // NostrEventSignedOpened to V2
  test("should convert validNostrEventSignedOpened to Spasm", () => {
    const input = JSON.parse(JSON.stringify(validNostrEventSignedOpened));
    const output = JSON.parse(JSON.stringify(validNostrEventSignedOpenedConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // NostrEventSignedOpened to V2
  test("should return null if validNostrEventSignedOpened has invalid signature", () => {
    const inputValid = {
      ...JSON.parse(JSON.stringify(validNostrEventSignedOpened)),
      sig: "908a15e46fb4d8675bab026fc230a0e3542bfade63da02d542fb78b2a8513fcd0092619a2c8c1221e581946e0191f2af505dfdf8657a414dbca329186f009262"
    };
    const inputInvalid1 = {
      ...JSON.parse(JSON.stringify(validNostrEventSignedOpened)),
      sig: "908a15e46fb4d8675bab026fc230a0e3542bfade63da02d542fb78b2a8513fcd0092619a2c8c1221e581946e0191f2af505dfdf8657a414dbca329186f009263"
    };
    const inputInvalid2 = {
      ...JSON.parse(JSON.stringify(validNostrEventSignedOpened)),
      sig: "nostr-sig-123"
    };
    const outputValid = JSON.parse(JSON.stringify(validNostrEventSignedOpenedConvertedToSpasmV2));
    // const output = null;
    expect(convertToSpasm(inputValid)).toEqual(outputValid);
    expect(convertToSpasm(inputInvalid1)).toEqual(null);
    expect(convertToSpasm(inputInvalid2)).toEqual(null);
  });

  // NostrSpasmEventSignedOpened to V2
  test("should convert validNostrSpasmEventSignedOpened to Spasm", () => {
    const input = JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpened));
    const output = JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpenedConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // NostrSpasmEventSignedOpened to V2
  test("should return null if validNostrSpasmEventSignedOpened has invalid signature", () => {
    const inputValid = {
      ...JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpened)),
      sig: "db60516accfc025582bf556e3c7660c89e3982d2a656201aaea4189c6d3e3779b202c60302e55ad782ca711df20550384516abe4d7387470bc83ac757ed8f0f1"
    };
    const inputInvalid1 = {
      ...JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpened)),
      sig: "db60516accfc025582bf556e3c7660c89e3982d2a656201aaea4189c6d3e3779b202c60302e55ad782ca711df20550384516abe4d7387470bc83ac757ed8f0f2"
    };
    const inputInvalid2 = {
      ...JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpened)),
      sig: "nostr-spasm-sig-123"
    };
    const outputValid = JSON.parse(JSON.stringify(validNostrSpasmEventSignedOpenedConvertedToSpasmV2));
    expect(convertToSpasm(inputValid)).toEqual(outputValid);
    expect(convertToSpasm(inputInvalid1)).toEqual(null);
    expect(convertToSpasm(inputInvalid2)).toEqual(null);
  });

  // PostWithNostrEventSignedOpened - old name
  // SpasmNostrEventSignedOpenedV0 - new name
  test("should convert validSpasmNostrEventSignedOpenedV0 to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validSpasmNostrEventSignedOpenedV0));
    const output = JSON.parse(JSON.stringify(validSpasmNostrEventSignedOpenedV0ConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // PostWithNostrSpasmEventSignedOpened - old name
  // SpasmNostrSpasmEventSignedOpenedV0 - new name
  test("should convert SpasmNostrSpasmEventSignedOpenedV0 to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validSpasmNostrSpasmEventSignedOpenedV0));
    const output = JSON.parse(JSON.stringify(validSpasmNostrSpasmEventSignedOpenedV0ConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  // validNostrReplyToDmpEvent
  // validNostrReplyToDmpEvent ("To the SPASM!") is tested as a
  // child (reply) of a validPostWithDmpEventSignedClosed inside
  // of SpasmDmpEventSignedClosedV0ConvertedToSpasmV2.
});

// convertToSpasm() for RSS items
// validPostWithRssItem - old name
// validSpasmEventRssItemV0 - new name
// validSpasmEventRssItemV0ConvertedToSpasmV2
describe("convertToSpasm() tests for RSS items", () => {
  // RssItem
  test("should convert validSpasmEventRssItemV0 to Spasm", () => {
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2));
    expect(convertToSpasm(input)).toEqual(output);
  });
});

describe("convertToSpasm() tests for events with sharedBy", () => {
  test("convertToSpasm() tests for events with and without sharedBy", () => {
    const dmpConverted: SpasmEventV2 =
      JSON.parse(JSON.stringify(
        validDmpEventSignedOpenedConvertedToSpasmV2
    ));
    const dmpConvertedWithOneSharedBy: SpasmEventV2 = {
      ...dmpConverted,
      sharedBy: {
        ids: [ { value: "0x12345" } ]
      }
    }
    const dmpConvertedWithTwoSharedBy: SpasmEventV2 = {
      ...dmpConverted,
      sharedBy: {
        ids: [ { value: "0x12345" }, { value: "0x67890" } ]
      }
    }
    expect(convertToSpasm(copyOf(dmpConverted)))
      .toStrictEqual(copyOf(dmpConverted));
    expect(convertToSpasm(copyOf(dmpConvertedWithOneSharedBy)))
      .toStrictEqual(copyOf(dmpConvertedWithOneSharedBy));
    expect(convertToSpasm(copyOf(dmpConvertedWithTwoSharedBy)))
      .toStrictEqual(copyOf(dmpConvertedWithTwoSharedBy));
  });
});

describe("convertToSpasm() tests for events with parent and root events", () => {
  test("convertToSpasm() tests for converting events with and without parent and root events", () => {
    // const spasmReply: SpasmEventV2 =
    //   copyOf(validPostWithNostrReplyToDmpEvent)
    const spasmReplyWithSpasmParent: SpasmEventV2 =
      copyOf(validPostWithNostrReplyToDmpEventConvertedToSpasmV2WithSpasmParentEvent)
    expect(convertToSpasm(copyOf(spasmReplyWithSpasmParent)))
      .toStrictEqual(copyOf(spasmReplyWithSpasmParent));
  });
});

// convertToSpasm() with sanitizeEvent() with DOMPurify
// validPostWithRssItem - old name
// validSpasmEventRssItemV0 - new name
// validSpasmEventRssItemV0ConvertedToSpasmV2
describe("convertToSpasm() tests with sanitizeEvent", () => {
  // RssItem
  test("should return null if an event has malicious code", () => {
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const inputMalicious = {
      ...input,
      tags: ["dark", "<img src=x onerror=alert(1)//>"]
    }
    /* const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2)); */
    expect(convertToSpasm(inputMalicious)).toEqual(null);
  });
  test("should return null if an event has deeply nested malicious code", () => {
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const inputMalicious = {
      ...input,
      tags: [
        "dark", "forest",
        {
          object: {
            array: [ 
              [ 1, 2, "three", "<svg><g/onload=alert(2)//<p>" ]
            ]
          }
        }
      ]
    }
    /* const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2)); */
    expect(convertToSpasm(inputMalicious)).toEqual(null);
  });
  test("should return event with malicious input if xss sanitization is turned off", () => {
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const inputMalicious = {
      ...input,
      tags: ["dark", "<img src=x onerror=alert(1)//>"]
    }
    const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2));
    const outputMalicious = {
      ...output
    }
    outputMalicious.ids[0].value = "spasmid019c82c9d5bee650a096e15ddbe4567928c17343653f2a34f2f6d2dfc2f6c744d1"
    outputMalicious.keywords = ["dark", "<img src=x onerror=alert(1)//>", "cookies"]
    outputMalicious.siblings[0].originalObject.tags = ["dark", "<img src=x onerror=alert(1)//>"]
    expect(
      convertToSpasm(
        inputMalicious,
        { xss: { enableSanitization: false } }
      )
    ).toEqual(outputMalicious);
  });
  test("should return null if an event has deeply nested malicious code", () => {
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const inputMalicious = {
      ...input,
      tags: [
        "dark", "forest",
        {
          object: {
            array: [ 
              [ 1, 2, "three", "<svg><g/onload=alert(2)//<p>" ]
            ]
          }
        }
      ]
    }
    /* const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2)); */
    expect(convertToSpasm(inputMalicious)).toEqual(null);
  });
  test("should return valid spasm event if custom sanitization function returns unchanged value", () => {
    const customFunction = (value: any) => value
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2));
    expect(
      convertToSpasm(
        input,
        {
          xss: {
            sanitizationConfig: {
              customFunction: customFunction
            }
          }
        }
      )
    ).toEqual(output);
  });
  test("should return null if custom sanitization function returns changed value", () => {
    const customFunction = (value: any) => value + " changed"
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    // const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2));
    expect(
      convertToSpasm(
        input,
        {
          xss: {
            sanitizationConfig: {
              customFunction: customFunction
            }
          }
        }
      )
    ).toEqual(null);
  });
  test("should return null if sanitization max recursion depth exceeded", () => {
    // Hide console errors for invalid addresses during tests
    jest.spyOn(console, 'error').mockImplementation(() => {});
    const input = JSON.parse(JSON.stringify(validSpasmEventRssItemV0));
    // const output = JSON.parse(JSON.stringify(validSpasmEventRssItemV0ConvertedToSpasmV2));
    expect(
      convertToSpasm(
        input,
        {
          xss: {
            sanitizationConfig: {
              maxDepth: 0
            }
          }
        }
      )
    ).toEqual(null);
    // Restore console errors
    jest.restoreAllMocks();
  });
});

// convertToSpasm() for events with special chars
describe("convertToSpasm() tests for events with special characters", () => {
  test("should convert validPostWithRssItemSpecialChars to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validPostWithRssItemSpecialChars));
    const output = JSON.parse(JSON.stringify(validPostWithRssItemSpecialCharsConvertedToSpasmEventV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  test("should convert validPostWithRssItemTitleHasSpecialChars to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validPostWithRssItemTitleHasSpecialChars));
    const output = JSON.parse(JSON.stringify(validPostWithRssItemTitleHasSpecialCharsConvertedToSpasmEventV2));
    expect(convertToSpasm(input)).toEqual(output);
  });

  test("an attempt to convert validNostrSpasmEventSpasmV0WithInvalidHtmlTags to SpasmEventV2 should return null", () => {
    // Hide console errors for invalid addresses during tests
    jest.spyOn(console, 'error').mockImplementation(() => {});
    const input = JSON.parse(JSON.stringify(validNostrSpasmEventSpasmV0WithInvalidHtmlTags));
    // null because cannot identify event
    expect(convertToSpasm(input)).toEqual(null);
    // Restore console errors
    jest.restoreAllMocks();
  });

  test("should convert validRssItemWithEmoji to SpasmEventV2", () => {
    const input = JSON.parse(JSON.stringify(validRssItemWithEmoji));
    const output = JSON.parse(JSON.stringify(validRssItemWithEmojiConvertedToSpasmEvent2));
    expect(convertToSpasm(input)).toEqual(output);
  });
});

describe("convertToSpasm() tests for SpasmEventEnvelopeV2", () => {
  const convertConfigV2: CustomConvertToSpasmConfig = {to: {spasm: {version:"2.0.0"}}}
  const testDoubleConvert = (from: any, to: any, equal: boolean = true) => {
    const event = JSON.parse(JSON.stringify(from));
    const spasmEnvelope = convertToSpasmEventEnvelope(event, "2.0.0");
    const input = convertToSpasm(spasmEnvelope as UnknownEventV2, convertConfigV2);
    const output = JSON.parse(JSON.stringify(to));
    if (equal) { 
      expect(input).toEqual(output);
    } else {
      expect(input).not.toEqual(output);
    }
  };
  const testDoubleConvertWithTree = (from: any, to: any, equal: boolean = true) => {
    const event = JSON.parse(JSON.stringify(from));
    const spasmEnvelope = convertToSpasmEventEnvelopeWithTree(event, "2.0.0");
    const input = convertToSpasm(spasmEnvelope as UnknownEventV2, convertConfigV2);
    const output = JSON.parse(JSON.stringify(to));
    if (equal) { 
      expect(input).toEqual(output);
    } else {
      expect(input).not.toEqual(output);
    }
  };
  const testDoubleConvertDatabase = (from: any, to: any, equal: boolean = true) => {
    const event = JSON.parse(JSON.stringify(from));
    const spasmDatabase = convertToSpasmEventDatabase(event, "2.0.0");
    const input = convertToSpasm(spasmDatabase as UnknownEventV2, convertConfigV2);
    const output = JSON.parse(JSON.stringify(to));
    if (equal) { 
      expect(input).toEqual(output);
    } else {
      expect(input).not.toEqual(output);
    }
  };

  test("validDmpEvent, validDmpEvent, false", () => {
    testDoubleConvert(validDmpEvent, validDmpEvent, false)
  });

  test("validDmpEvent, convertToSpasm(validDmpEvent, convertConfigV2)", () => {
    testDoubleConvert(validDmpEvent, convertToSpasm(copyOf(validDmpEvent), convertConfigV2))
  });

  test("validDmpEvent, validDmpEventConvertedToSpasmEventV2", () => {
    testDoubleConvert(validDmpEvent, validDmpEventConvertedToSpasmEventV2)
  });

  test("validDmpEvent, validDmpEventSignedClosed, false", () => {
    testDoubleConvert(validDmpEvent, validDmpEventSignedClosed, false)
  });

  test("validDmpEvent, validDmpEventSignedClosedConvertedToSpasmV2, false", () => {
    testDoubleConvert(validDmpEvent, validDmpEventSignedClosedConvertedToSpasmV2, false)
  });

  test("validDmpEventSignedClosed, validDmpEventSignedClosed, false", () => {
    testDoubleConvert(validDmpEventSignedClosed, validDmpEventSignedClosed, false)
  });

  test("validDmpEventSignedClosed, convertToSpasm(validDmpEventSignedClosed, convertConfigV2)", () => {
    testDoubleConvert(validDmpEventSignedClosed, convertToSpasm(copyOf(validDmpEventSignedClosed), convertConfigV2))
  });

  test("validDmpEventSignedClosed, validDmpEventSignedClosedConvertedToSpasmV2", () => {
    testDoubleConvert(validDmpEventSignedClosed, validDmpEventSignedClosedConvertedToSpasmV2)
  });

  test("validDmpEventSignedOpened, validDmpEventSignedOpened, false", () => {
    testDoubleConvert(validDmpEventSignedOpened, validDmpEventSignedOpened, false)
  });

  test("validDmpEventSignedOpened, convertToSpasm(validDmpEventSignedOpened, convertConfigV2)", () => {
    testDoubleConvert(validDmpEventSignedOpened, convertToSpasm(copyOf(validDmpEventSignedOpened), convertConfigV2))
  });

  test("validDmpEventSignedOpened, validDmpEventSignedOpenedConvertedToSpasmV2", () => {
    testDoubleConvert(validDmpEventSignedOpened, validDmpEventSignedOpenedConvertedToSpasmV2)
  });

  test("validDmpEventSignedClosedConvertedToSpasmV2, validDmpEventSignedClosedConvertedToSpasmV2", () => {
    testDoubleConvert(validDmpEventSignedClosedConvertedToSpasmV2, validDmpEventSignedClosedConvertedToSpasmV2)
  });

  test("validDmpEventSignedOpenedConvertedToSpasmV2, validDmpEventSignedClosedConvertedToSpasmV2", () => {
    testDoubleConvert(validDmpEventSignedOpenedConvertedToSpasmV2, validDmpEventSignedClosedConvertedToSpasmV2)
  });

  test("multiple convertToSpasm for DmpEvent", () => {
    testDoubleConvert(
      copyOf(validDmpEventSignedOpenedConvertedToSpasmV2),
      convertToSpasm(
        convertToSpasm(
          convertToSpasm(
            copyOf(validDmpEventSignedClosedConvertedToSpasmV2),
            convertConfigV2
          ) as UnknownEventV2,
          convertConfigV2
        ) as UnknownEventV2,
        convertConfigV2
      )
    )
  });

  test("validPostWithDmpEventSignedClosedConvertedToSpasmEventEnvelopeV2, rest1", () => {
    const { children, ...rest1 } = validPostWithDmpEventSignedClosedConvertedToSpasmV2
    testDoubleConvert(validPostWithDmpEventSignedClosedConvertedToSpasmEventEnvelopeV2, rest1);
  });

  test("validDmpEventSignedClosedWithInvalidSigner, null", () => {
    testDoubleConvert(validDmpEventSignedClosedWithInvalidSigner, null)
  });

  test("validDmpEventSignedClosedWithInvalidSignature, null", () => {
    testDoubleConvert(validDmpEventSignedClosedWithInvalidSignature, null)
  });

  test("validDmpEventSignedClosedWithInvalidSignedString, null", () => {
    testDoubleConvert(validDmpEventSignedClosedWithInvalidSignedString, null)
  });

  test("should convert various Nostr events to SpasmEventEnvelopeV2 and then to SpasmEventV2", () => {
    testDoubleConvert(validNostrEvent, validNostrEvent, false);
    testDoubleConvert(validNostrEvent, validNostrEventConvertedToSpasmV2);
    testDoubleConvert(validNostrEventSignedOpened, validNostrEventSignedOpened, false);
    testDoubleConvert(validNostrEventSignedOpened, validNostrEventConvertedToSpasmV2, false);
    testDoubleConvert(validNostrEventSignedOpened, validNostrEventSignedOpenedConvertedToSpasmV2);
    testDoubleConvert(validNostrSpasmEvent, validNostrSpasmEventSignedOpened, false);
    testDoubleConvert(validNostrSpasmEvent, validNostrSpasmEventSignedOpenedConvertedToSpasmV2, false);
    testDoubleConvert(validNostrSpasmEvent, validNostrSpasmEventConvertedToSpasmV2);
    testDoubleConvert(validNostrSpasmEventSignedOpened, validNostrSpasmEventSignedOpened, false);
    testDoubleConvert(validNostrSpasmEventSignedOpened, validNostrSpasmEventSignedOpenedConvertedToSpasmV2);
    testDoubleConvert(
      copyOf(validNostrSpasmEventSignedOpened),
      convertToSpasm(
        convertToSpasm(
          convertToSpasm(
            copyOf(validNostrSpasmEventSignedOpened),
            convertConfigV2
          ) as UnknownEventV2,
          convertConfigV2
        ) as UnknownEventV2,
        convertConfigV2
      )
    )
    testDoubleConvert(validNostrSpasmEventSignedOpenedWithInvalidSignature, null)
    testDoubleConvert(validNostrSpasmEventSignedOpenedWithInvalidSigner, null)
    testDoubleConvert(validNostrSpasmEventSignedOpenedWithInvalidContent, null)
  });

  test("should convert various RSS items to SpasmEventEnvelopeV2 and then to SpasmEventV2", () => {
    testDoubleConvert(validPostWithRssItemSpecialChars, validPostWithRssItemSpecialChars, false);
    testDoubleConvert(validPostWithRssItemSpecialChars, validPostWithRssItemSpecialCharsConvertedToSpasmEventV2);
    testDoubleConvert(validPostWithRssItemSpecialCharsConvertedToSpasmEventV2, validPostWithRssItemSpecialCharsConvertedToSpasmEventV2);

    testDoubleConvert(validNostrEvent, validNostrEventConvertedToSpasmV2);
  });

  test("converting various edge case events to SpasmEventEnvelopeV2 and then to SpasmEventV2", () => {
    // Hide console errors for invalid addresses during tests
    jest.spyOn(console, 'error').mockImplementation(() => {});
    testDoubleConvert(validNostrSpasmEventSpasmV0WithInvalidHtmlTags, null);
    // Restore console errors
    jest.restoreAllMocks();
  });

  test("converting events with comments to SpasmEventEnvelopeWithTreeV2 and then to SpasmEventV2", () => {
    testDoubleConvertWithTree(validPostWithDmpEventSignedClosed, validPostWithDmpEventSignedClosedConvertedToSpasmV2);
    testDoubleConvertWithTree([], null);
    testDoubleConvert(validPostWithDmpEventSignedClosed, validPostWithDmpEventSignedClosedConvertedToSpasmV2, false);
  });

  test("converting various events SpasmEventDatabaseV2 and then to SpasmEventV2", () => {
    testDoubleConvertDatabase(validDmpEventSignedClosed, validDmpEventSignedClosedConvertedToSpasmV2);
    testDoubleConvertDatabase(validDmpEventSignedClosed, validDmpEventSignedClosed, false);
    testDoubleConvertDatabase(
      validNostrSpasmEventSignedOpened,
      convertToSpasm(
        convertToSpasm(
          convertToSpasm(
            validNostrSpasmEventSignedOpened,
            convertConfigV2
          ) as UnknownEventV2,
          convertConfigV2
        ) as UnknownEventV2,
        convertConfigV2
      )
    )
    testDoubleConvertDatabase(validNostrSpasmEventSignedOpenedWithInvalidSignature, null)
    testDoubleConvertDatabase(validNostrSpasmEventSignedOpenedWithInvalidSigner, null)
    testDoubleConvertDatabase(validNostrSpasmEventSignedOpenedWithInvalidContent, null)
  });
});

describe("convertToSpasmEventEnvelopeWithTree() tests", () => {
  test("convertToSpasmEventEnvelopeWithTree() should convert event with relatives", () => {
    const envelopeWithTree =
      convertToSpasmEventEnvelopeWithTree(copyOf(
        validDmpEventSignedClosedConvertedToSpasmV2WithTwoChildren
      ), "2.0.0")
    expect(envelopeWithTree).not.toEqual(null)
    expect(envelopeWithTree?.type).toStrictEqual("SpasmEventEnvelopeWithTreeV2")
    expect(envelopeWithTree?.children?.length).toStrictEqual(2)
    expect(envelopeWithTree?.children?.[0]?.event?.type).toStrictEqual("SpasmEventEnvelopeWithTreeV2")
    expect(envelopeWithTree?.children?.[0]?.event?.ids).toStrictEqual(
      copyOf(validPostWithNostrReplyToDmpEventConvertedToSpasmV2).ids
    )
    expect(envelopeWithTree).toStrictEqual(copyOf(
      validDmpEventSignedClosedConvertedToSpasmEventEnvelopeWithTreeV2WithTwoChildren
    ))

    // Convert envelopeWithTree back to SpasmEvent
    const convertConfigV2: CustomConvertToSpasmConfig =
      {to: {spasm: {version:"2.0.0"}}}
    const envelopeWithTreeConvertedToSpasmEventV2 =
      convertToSpasm(copyOf(envelopeWithTree), convertConfigV2)
    const spasmEvent = envelopeWithTreeConvertedToSpasmEventV2
    expect(spasmEvent).not.toEqual(null)
    expect(spasmEvent?.type).toStrictEqual("SpasmEventV2");
    expect(spasmEvent?.title).toStrictEqual("genesis");
    expect("children" in spasmEvent!).toStrictEqual(true)
    expect(spasmEvent?.children?.length).toStrictEqual(2);
    expect(spasmEvent?.children?.[0].event?.type)
      .toStrictEqual("SpasmEventV2");
    expect(spasmEvent?.children?.[1].event?.type)
      .toStrictEqual("SpasmEventV2");
    expect(spasmEvent?.children?.[0].event).toStrictEqual(
      validPostWithNostrReplyToDmpEventConvertedToSpasmV2
    );

    // Testing max recursion with depth 0
    const envelopeWithTreeMaxDepth0 =
      convertToSpasmEventEnvelopeWithTree(copyOf(
        validDmpEventSignedClosedConvertedToSpasmV2WithTwoChildren
      ), "2.0.0", 0, 0)
    expect(envelopeWithTreeMaxDepth0).not.toEqual(null)
    expect(envelopeWithTreeMaxDepth0?.type)
      .toStrictEqual("SpasmEventEnvelopeWithTreeV2")
    expect("title" in envelopeWithTreeMaxDepth0!)
      .toStrictEqual(false)
    expect("content" in envelopeWithTreeMaxDepth0!)
      .toStrictEqual(false)
    expect("children" in envelopeWithTreeMaxDepth0!)
      .toStrictEqual(false)

    // Testing max recursion with depth 1
    const envelopeWithTreeMaxDepth1 =
      convertToSpasmEventEnvelopeWithTree(copyOf(
        validDmpEventSignedClosedConvertedToSpasmV2WithTwoChildren
      ), "2.0.0", 0, 1)
    expect(envelopeWithTreeMaxDepth1).not.toEqual(null)
    expect(envelopeWithTreeMaxDepth1?.type)
      .toStrictEqual("SpasmEventEnvelopeWithTreeV2")
    expect("title" in envelopeWithTreeMaxDepth1!)
      .toStrictEqual(false)
    expect("content" in envelopeWithTreeMaxDepth1!)
      .toStrictEqual(false)
    expect("children" in envelopeWithTreeMaxDepth1!)
      .toStrictEqual(true)
    expect(envelopeWithTreeMaxDepth1?.children?.[0]?.event?.type).toStrictEqual("SpasmEventEnvelopeWithTreeV2")
    expect(envelopeWithTreeMaxDepth1?.children?.[0]?.event?.ids).toStrictEqual(
      copyOf(validPostWithNostrReplyToDmpEventConvertedToSpasmV2).ids
    )
  });
});

describe("convertToSpasmEventEnvelope() tests", () => {
  test("convertToSpasmEventEnvelope() should convert event with relatives", () => {
    const envelope = convertToSpasmEventEnvelope(copyOf(
      validDmpEventSignedClosedConvertedToSpasmV2WithTwoChildren
    ), "2.0.0")
    expect(envelope).not.toEqual(null)
    expect(envelope!.type).toStrictEqual("SpasmEventEnvelopeV2")
    expect("title" in envelope!).toStrictEqual(false)
    expect("content" in envelope!).toStrictEqual(false)
    expect("children" in envelope!).toStrictEqual(false)
    expect(envelope?.ids).toStrictEqual(
      copyOf(validDmpEventSignedClosedConvertedToSpasmV2.ids)
    )

    // Convert envelope back to SpasmEvent
    const convertConfigV2: CustomConvertToSpasmConfig =
      {to: {spasm: {version:"2.0.0"}}}
    const envelopeConvertedToSpasmEventV2 =
      convertToSpasm(copyOf(envelope), convertConfigV2)
    const spasmEvent = envelopeConvertedToSpasmEventV2
    expect(spasmEvent).not.toEqual(null)
    expect(spasmEvent?.type).toStrictEqual("SpasmEventV2");
    expect(spasmEvent?.title).toStrictEqual("genesis");
    expect("children" in spasmEvent!).toStrictEqual(false)
  });
});
